Correct numeric value format

+++ src/prebuilt/ver1_46/background.rs
@@ -10,11 +10,11 @@ pub struct Background {}
 impl Background {
     pub fn template() -> Token {
         let mut root = LinkedHashMap::with_capacity(6);
-        root.push_back("ascii_filename", DynString::with_capacity(0 as u16, ""));
-        root.push_back("picture_version", 3 as u32);
-        root.push_back("bitmap_width", 0 as u32);
-        root.push_back("bitmap_height", 0 as u32);
-        root.push_back("picture_orientation", 0 as i16);
+        root.push_back("ascii_filename", DynString::with_capacity(0_u16, ""));
+        root.push_back("picture_version", 3_u32);
+        root.push_back("bitmap_width", 0_u32);
+        root.push_back("bitmap_height", 0_u32);
+        root.push_back("picture_orientation", 0_i16);
 
         root.push_back("bitmap_info", BitmapInfo::template());
         root.patchs.insert(
