+++ src/prebuilt/ver1_46/cinematics.rs
@@ -8,9 +8,9 @@ pub struct Cinematics {}
 impl Cinematics {
     pub fn template() -> Token {
         let mut root = LinkedHashMap::with_capacity(3);
-        root.push_back("ascii_pregame", DynString::with_capacity(0 as u16, ""));
-        root.push_back("ascii_victory", DynString::with_capacity(0 as u16, ""));
-        root.push_back("ascii_loss", DynString::with_capacity(0 as u16, ""));
+        root.push_back("ascii_pregame", DynString::with_capacity(0_u16, ""));
+        root.push_back("ascii_victory", DynString::with_capacity(0_u16, ""));
+        root.push_back("ascii_loss", DynString::with_capacity(0_u16, ""));
 
         root.into()
     }
