+++ src/prebuilt/ver1_46/player_units.rs
@@ -7,7 +7,7 @@ pub struct PlayerUnits {}
 impl PlayerUnits {
     pub fn template() -> Token {
         let mut root = LinkedHashMap::with_capacity(2);
-        root.push_back("unit_count", 0 as u32);
+        root.push_back("unit_count", 0_u32);
         root.push_back("units", vec![Unit::template(); 1]);
         root.patchs.insert(
             "units".to_string(),
